+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Support `osls` v4 / frameworks that removed the bundled AWS SDK v2 module. The plugin no longer accesses `providers.aws.sdk` or `providers.aws.getCredentials()` when `getAwsSdkV3Config()` is available, avoiding the `AWS_SDK_V2_SURFACE_REMOVED` error. AWS SDK v3 clients resolve custom endpoints and credentials natively on those frameworks ([678](https://github.com/amplify-education/serverless-domain-manager/issues/678))
+
 ## [10.2.0] - 2026-06-09
 
 ### Changed